Update test scenes to actually cover submission logic
